@@ -1,24 +1,34 @@
-import { LanguageCode, SourceInfo, TagType } from "paperback-extensions-common";
-import { MangaStream } from '../MangaStream'
+/* eslint-disable linebreak-style */
+import {
+    LanguageCode,
+    SourceInfo,
+    ContentRating,
+    TagType
+} from 'paperback-extensions-common'
 
-const MANGAGENKI_DOMAIN = "https://mangagenki.com"
+import {
+    MangaStream,
+    getExportVersion
+} from '../MangaStream'
+
+const MANGAGENKI_DOMAIN = 'https://mangagenki.com'
 
 export const MangaGenkiInfo: SourceInfo = {
-    version: '1.0.2',
+    version: getExportVersion('0.0.0'),
     name: 'MangaGenki',
     description: 'Extension that pulls manga from MangaGenki',
     author: 'Netsky',
     authorWebsite: 'http://github.com/TheNetsky',
-    icon: "icon.png",
-    hentaiSource: false,
+    icon: 'icon.png',
+    contentRating: ContentRating.ADULT,
     websiteBaseURL: MANGAGENKI_DOMAIN,
     sourceTags: [
         {
-            text: "Notifications",
+            text: 'Notifications',
             type: TagType.GREEN
         },
         {
-            text: "18+",
+            text: '18+',
             type: TagType.YELLOW
         }
     ]
@@ -29,7 +39,6 @@ export class MangaGenki extends MangaStream {
 
     baseUrl: string = MANGAGENKI_DOMAIN
     languageCode: LanguageCode = LanguageCode.ENGLISH
-    hasAdvancedSearchPage: boolean = true
 
     //----MANGA DETAILS SELECTORS
     /*
@@ -48,16 +57,16 @@ export class MangaGenki extends MangaStream {
     //Disabling some of these will cause some Home-Page tests to fail, edit these test files to match the setting.
     //Always be sure to test this in the app!
 
-    homescreen_PopularToday_enabled: boolean = true
+    override homescreen_PopularToday_enabled = true
 
-    homescreen_LatestUpdate_enabled: boolean = true
+    override homescreen_LatestUpdate_enabled = true
 
-    homescreen_NewManga_enabled: boolean = true
-    homescreen_NewManga_selector: string = "h3:contains(New Titles)"
+    override homescreen_NewManga_enabled = true
+    override homescreen_NewManga_selector = 'h3:contains(New Titles)'
 
-    homescreen_TopAllTime_enabled: boolean = true
-    homescreen_TopMonthly_enabled: boolean = true
-    homescreen_TopWeekly_enabled: boolean = true
+    override homescreen_TopAllTime_enabled = true
+    override homescreen_TopMonthly_enabled = true
+    override homescreen_TopWeekly_enabled = true
 
     /*
     ----TAG SELECTORS
