@@ -1,3 +1,4 @@
+/* eslint-disable linebreak-style */
 import {
     Chapter,
     ChapterDetails,
@@ -5,13 +6,17 @@ import {
     LanguageCode,
     Manga,
     MangaUpdates,
-    PagedResults, RequestHeaders,
+    PagedResults,
     SearchRequest,
-    Source, TagSection,
+    Source,
+    TagSection,
 
-} from "paperback-extensions-common"
+} from 'paperback-extensions-common'
 
-import { MangaStreamParser, UpdatedManga } from './MangaStreamParser'
+import {
+    MangaStreamParser,
+    UpdatedManga
+} from './MangaStreamParser'
 
 interface TimeAgo {
     now: string[],
@@ -23,7 +28,7 @@ interface TimeAgo {
     hours: string[],
     minutes: string[],
     seconds: string[]
-};
+}
 interface dateMonths {
     january: string,
     february: string,
@@ -37,11 +42,18 @@ interface dateMonths {
     october: string,
     november: string,
     december: string
-};
+}
 interface StatusTypes {
     ONGOING: string,
     COMPLETED: string
-};
+}
+
+
+// Set the version for the base, changing this version will change the versions of all sources
+const BASE_VERSION = '2.0.2'
+export const getExportVersion = (EXTENSION_VERSION: string): string => {
+    return BASE_VERSION.split('.').map((x, index) => Number(x) + Number(EXTENSION_VERSION.split('.')[index])).join('.')
+}
 
 export abstract class MangaStream extends Source {
     /**
@@ -57,7 +69,7 @@ export abstract class MangaStream extends Source {
     /**
      * Helps with CloudFlare for some sources, makes it worse for others; override with empty string if the latter is true
      */
-    userAgentRandomizer: string = `Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:77.0) Gecko/20100101 Firefox/78.0${Math.floor(Math.random() * 100000)}`
+    userAgentRandomizer = `Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:77.0) Gecko/20100101 Firefox/78.0${Math.floor(Math.random() * 100000)}`
 
     //----GENERAL SELECTORS----
     /**
@@ -65,18 +77,13 @@ export abstract class MangaStream extends Source {
      * Eg. https://mangadark.com/manga/mashle-magic-and-muscles the pathname would be "manga"
      * Default = "manga"
      */
-    sourceTraversalPathName: string = "manga"
-
-    /**
-     * N/A.
-     */
-    hasAdvancedSearchPage: boolean = false
+    sourceTraversalPathName = 'manga'
 
     /**
      * Fallback image if no image is present
      * Default = "https://i.imgur.com/GYUxEX8.png"
      */
-    fallbackImage: string = "https://i.imgur.com/GYUxEX8.png"
+    fallbackImage = 'https://i.imgur.com/GYUxEX8.png'
 
     //----MANGA DETAILS SELECTORS----
     /**
@@ -85,35 +92,35 @@ export abstract class MangaStream extends Source {
      * Leave default if not used!
      * Default = "b:contains(Alternative Titles)"
     */
-    manga_selector_AlternativeTitles: string = "Alternative Titles"
+    manga_selector_AlternativeTitles = 'Alternative Titles'
     /**
      * The selector for authors.
      * This can change depending on the language
      * Leave default if not used!
      * Default = "Author" (English)
     */
-    manga_selector_author: string = "Author"
+    manga_selector_author = 'Author'
     /**
      * The selector for artists.
      * This can change depending on the language
      * Leave default if not used!
      * Default = "Artist" (English)
     */
-    manga_selector_artist: string = "Artist"
+    manga_selector_artist = 'Artist'
 
-    manga_selector_status: string = "Status"
+    manga_selector_status = 'Status'
 
-    manga_tag_selector_box: string = "span.mgen"
+    manga_tag_selector_box = 'span.mgen'
 
-    manga_tag_TraversalPathName: string = "genres"
+    manga_tag_TraversalPathName = 'genres'
     /**
      * The selector for the manga status.
      * These can change depending on the language
      * Default = "ONGOING: "ONGOING", COMPLETED: "COMPLETED"
     */
     manga_StatusTypes: StatusTypes = {
-        ONGOING: "ONGOING",
-        COMPLETED: "COMPLETED"
+        ONGOING: 'ONGOING',
+        COMPLETED: 'COMPLETED'
     }
 
 
@@ -123,18 +130,18 @@ export abstract class MangaStream extends Source {
      * Default = English Translation
      */
     dateMonths: dateMonths = {
-        january: "January",
-        february: "February",
-        march: "March",
-        april: "April",
-        may: "May",
-        june: "June",
-        july: "July",
-        august: "August",
-        september: "September",
-        october: "October",
-        november: "November",
-        december: "December"
+        january: 'January',
+        february: 'February',
+        march: 'March',
+        april: 'April',
+        may: 'May',
+        june: 'June',
+        july: 'July',
+        august: 'August',
+        september: 'September',
+        october: 'October',
+        november: 'November',
+        december: 'December'
     };
     /**
      * In this object, add the site's translations for the following time formats, case insensitive.
@@ -142,15 +149,15 @@ export abstract class MangaStream extends Source {
      * Default =  English Translation
      */
     dateTimeAgo: TimeAgo = {
-        now: ["less than an hour", "just now"],
-        yesterday: ["yesterday"],
-        years: ["year"],
-        months: ["month"],
-        weeks: ["week"],
-        days: ["day"],
-        hours: ["hour"],
-        minutes: ["min"],
-        seconds: ["second"]
+        now: ['less than an hour', 'just now'],
+        yesterday: ['yesterday'],
+        years: ['year'],
+        months: ['month'],
+        weeks: ['week'],
+        days: ['day'],
+        hours: ['hour'],
+        minutes: ['min'],
+        seconds: ['second']
     };
     //----CHAPTER SELECTORS----
     /**
@@ -158,13 +165,13 @@ export abstract class MangaStream extends Source {
      * This box contains all the chapter items
      * Default = "div#chapterlist.eplister"
     */
-    chapter_selector_box: string = "div#chapterlist"
+    chapter_selector_box = 'div#chapterlist'
     /**
      * The selector for each individual chapter element
      * This is the element for each small box containing the chapter information
      * Default = "li"
     */
-    chapter_selector_item: string = "li"
+    chapter_selector_item = 'li'
 
     //----TAGS SELECTORS----
     /**
@@ -172,23 +179,23 @@ export abstract class MangaStream extends Source {
      * Eg. https://mangadark.com/genres/ needs this selector to be set to "/genres/"
      * Default = ""
     */
-    tags_SubdirectoryPathName: string = ""
+    tags_SubdirectoryPathName = ''
     /**
      * The selector to select the box with all the genres
      * Default = "ul.genre"
     */
-    tags_selector_box: string = "ul.genre"
+    tags_selector_box = 'ul.genre'
     /**
      * The selector to select each individual genre box
      * Default = "li"
     */
-    tags_selector_item: string = "li"
+    tags_selector_item = 'li'
     /**
      * The selector to select the label name
      * Some sites have a result number after the genre name, this selector allows you to filter this.
      * Default = ""
     */
-    tags_selector_label: string = ""
+    tags_selector_label = ''
 
     //----HOMESCREEN SELECTORS----
     /**
@@ -197,123 +204,137 @@ export abstract class MangaStream extends Source {
      * Enabled Default = true
      * Selector Default = "h2:contains(Popular Today)"
     */
-    homescreen_PopularToday_enabled: boolean = true
-    homescreen_PopularToday_selector: string = "h2:contains(Popular Today)"
+    homescreen_PopularToday_enabled = true
+    homescreen_PopularToday_selector = 'h2:contains(Popular Today)'
 
-    homescreen_LatestUpdate_enabled: boolean = true
-    homescreen_LatestUpdate_selector_box: string = "h2:contains(Latest Update)"
-    homescreen_LatestUpdate_selector_item: string = "div.uta"
+    homescreen_LatestUpdate_enabled = true
+    homescreen_LatestUpdate_selector_box = 'h2:contains(Latest Update)'
+    homescreen_LatestUpdate_selector_item = 'div.uta'
 
-    homescreen_NewManga_enabled: boolean = true
-    homescreen_NewManga_selector: string = "h3:contains(New Series)"
+    homescreen_NewManga_enabled = true
+    homescreen_NewManga_selector = 'h3:contains(New Series)'
 
-    homescreen_TopAllTime_enabled: boolean = true
-    homescreen_TopAllTime_selector: string = "div.serieslist.pop.wpop.wpop-alltime"
+    homescreen_TopAllTime_enabled = true
+    homescreen_TopAllTime_selector = 'div.serieslist.pop.wpop.wpop-alltime'
 
-    homescreen_TopMonthly_enabled: boolean = true
-    homescreen_TopMonthly_selector: string = "div.serieslist.pop.wpop.wpop-monthly"
+    homescreen_TopMonthly_enabled = true
+    homescreen_TopMonthly_selector = 'div.serieslist.pop.wpop.wpop-monthly'
 
-    homescreen_TopWeekly_enabled: boolean = true
-    homescreen_TopWeekly_selector: string = "div.serieslist.pop.wpop.wpop-weekly"
+    homescreen_TopWeekly_enabled = true
+    homescreen_TopWeekly_selector = 'div.serieslist.pop.wpop.wpop-weekly'
 
     //----REQUEST MANAGER----
     requestManager = createRequestManager({
-        requestsPerSecond: 2.5,
+        requestsPerSecond: 3,
         requestTimeout: 15000,
     });
 
     parser = new MangaStreamParser();
 
-    getMangaShareUrl(mangaId: string): string {
-        return `${this.baseUrl}/${this.sourceTraversalPathName}/${mangaId}/`;
+
+    override getMangaShareUrl(mangaId: string): string {
+        return `${this.baseUrl}/${this.sourceTraversalPathName}/${mangaId}/`
     }
 
-    async getMangaDetails(mangaId: string): Promise<Manga> {
+    override async getMangaDetails(mangaId: string): Promise<Manga> {
         const request = createRequestObject({
             url: `${this.baseUrl}/${this.sourceTraversalPathName}/${mangaId}/`,
             method: 'GET',
             headers: this.constructHeaders({})
-        });
+        })
 
-        const response = await this.requestManager.schedule(request, 1);
-        this.CloudFlareError(response.status);
-        let $ = this.cheerio.load(response.data);
+        const response = await this.requestManager.schedule(request, 1)
+        this.CloudFlareError(response.status)
+        const $ = this.cheerio.load(response.data)
 
-        return this.parser.parseMangaDetails($, mangaId, this);
+        return this.parser.parseMangaDetails($, mangaId, this)
     }
 
-    async getChapters(mangaId: string): Promise<Chapter[]> {
+    override async getChapters(mangaId: string): Promise<Chapter[]> {
         const request = createRequestObject({
             url: `${this.baseUrl}/${this.sourceTraversalPathName}/${mangaId}/`,
             method: 'GET',
             headers: this.constructHeaders({})
-        });
+        })
 
-        const response = await this.requestManager.schedule(request, 1);
-        this.CloudFlareError(response.status);
-        let $ = this.cheerio.load(response.data);
+        const response = await this.requestManager.schedule(request, 1)
+        this.CloudFlareError(response.status)
+        const $ = this.cheerio.load(response.data)
 
-        return this.parser.parseChapterList($, mangaId, this);
+        return this.parser.parseChapterList($, mangaId, this)
     }
 
-    async getChapterDetails(mangaId: string, chapterId: string): Promise<ChapterDetails> {
+    override async getChapterDetails(mangaId: string, chapterId: string): Promise<ChapterDetails> {
         const request = createRequestObject({
             url: `${this.baseUrl}/${chapterId}/`,
             method: 'GET',
             headers: this.constructHeaders({}),
-        });
+        })
 
-        const response = await this.requestManager.schedule(request, 1);
-        this.CloudFlareError(response.status);
-        return this.parser.parseChapterDetails(response.data, mangaId, chapterId);
+        const response = await this.requestManager.schedule(request, 1)
+        this.CloudFlareError(response.status)
+        return this.parser.parseChapterDetails(response.data, mangaId, chapterId)
     }
 
-    async getTags(): Promise<TagSection[] | null> {
+    override async getTags(): Promise<TagSection[]> {
         const request = createRequestObject({
             url: `${this.baseUrl}/`,
-            method: "GET",
+            method: 'GET',
+            headers: this.constructHeaders({}),
             param: this.tags_SubdirectoryPathName
-        });
+        })
 
-        const response = await this.requestManager.schedule(request, 1);
-        this.CloudFlareError(response.status);
-        const $ = this.cheerio.load(response.data);;
-        return this.parser.parseTags($, this);
+        const response = await this.requestManager.schedule(request, 1)
+        this.CloudFlareError(response.status)
+        const $ = this.cheerio.load(response.data)
+        return this.parser.parseTags($, this)
     }
 
-    async searchRequest(query: SearchRequest, metadata: any): Promise<PagedResults> {
-        let page = metadata?.page ?? 0;
-        const search = this.parser.generateSearch(query);
-        const request = createRequestObject({
-            url: `${this.baseUrl}/page/${page}/?s=`,
-            method: "GET",
-            param: search
-        });
+    override async getSearchResults(query: SearchRequest, metadata: any): Promise<PagedResults> {
+        const page = metadata?.page ?? 1
+        let request
 
-        const response = await this.requestManager.schedule(request, 1);
-        this.CloudFlareError(response.status);
-        const $ = this.cheerio.load(response.data);
-        const manga = this.parser.parseSearchResults($, this);
-        metadata = !this.parser.isLastPage($, "search_request") ? { page: page + 1 } : undefined;
+        if (query.title) {
+            request = createRequestObject({
+                url: `${this.baseUrl}/page/${page}/?s=`,
+                method: 'GET',
+                headers: this.constructHeaders({}),
+                param: encodeURI(query.title)
+            })
+        } else {
+            request = createRequestObject({
+                url: `${this.baseUrl}/`,
+                method: 'GET',
+                headers: this.constructHeaders({}),
+                param: `genres/${query?.includedTags?.map((x: any) => x.id)[0]}/page/${page}`
+            })
+        }
+
+        const response = await this.requestManager.schedule(request, 1)
+        this.CloudFlareError(response.status)
+        const $ = this.cheerio.load(response.data)
+        const manga = this.parser.parseSearchResults($, this)
+        metadata = !this.parser.isLastPage($, 'search_request') ? { page: page + 1 } : undefined
 
         return createPagedResults({
             results: manga,
             metadata
-        });
+        })
     }
 
-    async filterUpdatedManga(mangaUpdatesFoundCallback: (updates: MangaUpdates) => void, time: Date, ids: string[]): Promise<void> {
-        let page = 1;
+    override async filterUpdatedManga(mangaUpdatesFoundCallback: (updates: MangaUpdates) => void, time: Date, ids: string[]): Promise<void> {
+        let page = 1
         let updatedManga: UpdatedManga = {
             ids: [],
             loadMore: true,
-        };
+        }
 
         while (updatedManga.loadMore) {
             const request = createRequestObject({
                 url: `${this.baseUrl}/page/${page++}/`,
-                method: "GET",
-            });
+                method: 'GET',
+                headers: this.constructHeaders({})
+            })
 
             const response = await this.requestManager.schedule(request, 1)
             const $ = this.cheerio.load(response.data)
@@ -322,74 +343,76 @@ export abstract class MangaStream extends Source {
             if (updatedManga.ids.length > 0) {
                 mangaUpdatesFoundCallback(createMangaUpdates({
                     ids: updatedManga.ids
-                }));
+                }))
             }
         }
 
     }
 
-    async getHomePageSections(sectionCallback: (section: HomeSection) => void): Promise<void> {
-        const section1 = createHomeSection({ id: 'popular_today', title: 'Popular Today', view_more: true });
-        const section2 = createHomeSection({ id: 'latest_update', title: 'Latest Updates', view_more: true });
-        const section3 = createHomeSection({ id: 'new_titles', title: 'New Titles', view_more: true });
-        const section4 = createHomeSection({ id: 'top_alltime', title: 'Top All Time', view_more: false });
-        const section5 = createHomeSection({ id: 'top_monthly', title: 'Top Monthly', view_more: false });
-        const section6 = createHomeSection({ id: 'top_weekly', title: 'Top Weekly', view_more: false });
+    override async getHomePageSections(sectionCallback: (section: HomeSection) => void): Promise<void> {
+        const section1 = createHomeSection({ id: 'popular_today', title: 'Popular Today', view_more: true })
+        const section2 = createHomeSection({ id: 'latest_update', title: 'Latest Updates', view_more: true })
+        const section3 = createHomeSection({ id: 'new_titles', title: 'New Titles', view_more: true })
+        const section4 = createHomeSection({ id: 'top_alltime', title: 'Top All Time', view_more: false })
+        const section5 = createHomeSection({ id: 'top_monthly', title: 'Top Monthly', view_more: false })
+        const section6 = createHomeSection({ id: 'top_weekly', title: 'Top Weekly', view_more: false })
 
-        const sections: any[] = [];
-        if (this.homescreen_PopularToday_enabled) sections.push(section1);
-        if (this.homescreen_LatestUpdate_enabled) sections.push(section2);
-        if (this.homescreen_NewManga_enabled) sections.push(section3);
-        if (this.homescreen_TopAllTime_enabled) sections.push(section4);
-        if (this.homescreen_TopMonthly_enabled) sections.push(section5);
-        if (this.homescreen_TopWeekly_enabled) sections.push(section6);
+        const sections: any[] = []
+        if (this.homescreen_PopularToday_enabled) sections.push(section1)
+        if (this.homescreen_LatestUpdate_enabled) sections.push(section2)
+        if (this.homescreen_NewManga_enabled) sections.push(section3)
+        if (this.homescreen_TopAllTime_enabled) sections.push(section4)
+        if (this.homescreen_TopMonthly_enabled) sections.push(section5)
+        if (this.homescreen_TopWeekly_enabled) sections.push(section6)
 
         const request = createRequestObject({
             url: `${this.baseUrl}/`,
-            method: "GET",
-        });
+            method: 'GET',
+            headers: this.constructHeaders({})
+        })
 
-        const response = await this.requestManager.schedule(request, 1);
-        this.CloudFlareError(response.status);
-        const $ = this.cheerio.load(response.data);
-        this.parser.parseHomeSections($, sections, sectionCallback, this);
+        const response = await this.requestManager.schedule(request, 1)
+        this.CloudFlareError(response.status)
+        const $ = this.cheerio.load(response.data)
+        this.parser.parseHomeSections($, sections, sectionCallback, this)
     }
 
-    async getViewMoreItems(homepageSectionId: string, metadata: any): Promise<PagedResults | null> {
-        let page: number = metadata?.page ?? 1;
-        let param = "";
+    override async getViewMoreItems(homepageSectionId: string, metadata: any): Promise<PagedResults> {
+        const page: number = metadata?.page ?? 1
+        let param = ''
         switch (homepageSectionId) {
-            case "new_titles":
-                param = `/${this.sourceTraversalPathName}/?page=${page}&order=latest`;
-                break;
-            case "latest_update":
-                param = `/${this.sourceTraversalPathName}/?page=${page}&order=update`;
-                break;
-            case "popular_today":
-                param = `/${this.sourceTraversalPathName}/?page=${page}&order=popular`;
-                break;
+            case 'new_titles':
+                param = `/${this.sourceTraversalPathName}/?page=${page}&order=latest`
+                break
+            case 'latest_update':
+                param = `/${this.sourceTraversalPathName}/?page=${page}&order=update`
+                break
+            case 'popular_today':
+                param = `/${this.sourceTraversalPathName}/?page=${page}&order=popular`
+                break
             default:
-                return Promise.resolve(null);;
+                throw new Error(`Invalid homeSectionId | ${homepageSectionId}`)
         }
 
         const request = createRequestObject({
             url: `${this.baseUrl}/`,
-            method: "GET",
+            method: 'GET',
+            headers: this.constructHeaders({}),
             param,
-        });
+        })
 
-        const response = await this.requestManager.schedule(request, 1);
-        const $ = this.cheerio.load(response.data);
+        const response = await this.requestManager.schedule(request, 1)
+        const $ = this.cheerio.load(response.data)
 
-        const manga = this.parser.parseViewMore($, this);
-        metadata = !this.parser.isLastPage($, "view_more") ? { page: page + 1 } : undefined;
+        const manga = this.parser.parseViewMore($, this)
+        metadata = !this.parser.isLastPage($, 'view_more') ? { page: page + 1 } : undefined
         return createPagedResults({
             results: manga,
             metadata
-        });
+        })
     }
 
-    getCloudflareBypassRequest() {
+    override getCloudflareBypassRequest() {
         return createRequestObject({
             url: `${this.baseUrl}/`,
             method: 'GET',
@@ -399,30 +422,15 @@ export abstract class MangaStream extends Source {
 
     constructHeaders(headers: any, refererPath?: string): any {
         if (this.userAgentRandomizer !== '') {
-            headers["user-agent"] = this.userAgentRandomizer;
+            headers['user-agent'] = this.userAgentRandomizer
         }
-        headers["referer"] = `${this.baseUrl}${refererPath ?? ''}`;
-        return headers;
-    }
-
-    globalRequestHeaders(): RequestHeaders {
-        if (this.userAgentRandomizer !== '') {
-            return {
-                "referer": `${this.baseUrl}/`,
-                "user-agent": this.userAgentRandomizer,
-                "accept": "image/jpeg,image/png,image/*;q=0.8"
-            }
-        } else {
-            return {
-                "referer": `${this.baseUrl}/`,
-                "accept": "image/jpeg,image/png,image/*;q=0.8"
-            }
-        }
+        headers['referer'] = `${this.baseUrl}${refererPath ?? ''}/`
+        return headers
     }
 
     CloudFlareError(status: any) {
         if (status == 503) {
-            throw new Error('CLOUDFLARE BYPASS ERROR:\nPlease go to Settings > Sources > \<\The name of this source\> and press Cloudflare Bypass');
+            throw new Error('CLOUDFLARE BYPASS ERROR:\nPlease go to Settings > Sources > <The name of this source> and press Cloudflare Bypass')
         }
     }
 
