@@ -2,7 +2,6 @@
 import {
     LanguageCode,
     SourceInfo,
-    TagType,
     ContentRating
 } from 'paperback-extensions-common'
 
@@ -11,29 +10,23 @@ import {
     getExportVersion
 } from '../MangaStream'
 
-const MANGADARK_DOMAIN = 'https://mangadark.com'
+const PMSCANS_DOMAIN = 'https://reader.pmscans.com'
 
-export const MangaDarkInfo: SourceInfo = {
+export const PMScansInfo: SourceInfo = {
     version: getExportVersion('0.0.0'),
-    name: 'MangaDark',
-    description: 'Extension that pulls manga from MangaDark',
+    name: 'PMScans',
+    description: 'Extension that pulls manga from PMScans',
     author: 'Netsky',
     authorWebsite: 'http://github.com/TheNetsky',
     icon: 'icon.png',
     contentRating: ContentRating.MATURE,
-    websiteBaseURL: MANGADARK_DOMAIN,
-    sourceTags: [
-        {
-            text: 'Notifications',
-            type: TagType.GREEN
-        }
-    ]
+    websiteBaseURL: PMSCANS_DOMAIN,
 }
 
-export class MangaDark extends MangaStream {
+export class PMScans extends MangaStream {
     //FOR ALL THE SELECTIONS, PLEASE CHECK THE MangaSteam.ts FILE!!!
 
-    baseUrl: string = MANGADARK_DOMAIN
+    baseUrl: string = PMSCANS_DOMAIN
     languageCode: LanguageCode = LanguageCode.ENGLISH
 
     //----MANGA DETAILS SELECTORS
@@ -48,15 +41,16 @@ export class MangaDark extends MangaStream {
     //    COMPLETED: "completed"
     //}
 
+
     //----HOMESCREEN SELECTORS
     //Disabling some of these will cause some Home-Page tests to fail, edit these test files to match the setting.
     //Always be sure to test this in the app!
 
     override homescreen_PopularToday_enabled = true
 
-    override homescreen_LatestUpdate_enabled = true
-
-    override homescreen_NewManga_enabled = true
+    override homescreen_LatestUpdate_enabled = false //Directly links to chapter, not detail page!
+    
+    override homescreen_NewManga_enabled = false
 
     override homescreen_TopAllTime_enabled = true
     override homescreen_TopMonthly_enabled = true
@@ -76,10 +70,5 @@ export class MangaDark extends MangaStream {
     tags_selector_item: string = "li"
     tags_selector_label: string = "span"
     */
-
-    override tags_SubdirectoryPathName = '/genres/'
-    override tags_selector_box = 'div.page'
-    override tags_selector_item = 'li'
-    override tags_selector_label = 'span'
 
 }
