@@ -4,14 +4,14 @@ import {
     SearchRequest,
 } from 'paperback-extensions-common'
 import { MangaStream } from '../MangaStream'
-import { eDoujin } from '../eDoujin/eDoujin'
+import { AsuraScans } from '../AsuraScans/AsuraScans'
 import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 
-describe('eDoujin Tests', () => {
+describe('AsuraScans Tests', () => {
 
     const wrapper: APIWrapper = new APIWrapper()
-    const source: MangaStream = new eDoujin(cheerio)
+    const source: MangaStream = new AsuraScans(cheerio)
     const expect = chai.expect
     chai.use(chaiAsPromised)
 
@@ -20,7 +20,7 @@ describe('eDoujin Tests', () => {
    * Try to choose a manga which is updated frequently, so that the historical checking test can
    * return proper results, as it is limited to searching 30 days back due to extremely long processing times otherwise.
    */
-    const mangaId = 'reformation-of-the-deadbeat-noble' // secret-class
+    const mangaId = 'legend-of-asura-the-venom-dragon' // secret-class
 
     it('Retrieve Manga Details', async () => {
         const details = await wrapper.getMangaDetails(source, mangaId)
@@ -67,7 +67,7 @@ describe('eDoujin Tests', () => {
     })
 
     it('Testing home page results for popular titles', async () => {
-        const results = await wrapper.getViewMoreItems(source, 'popular_today', {}, 1)
+        const results = await wrapper.getViewMoreItems(source, 'latest_update', {}, 1)
 
         expect(results, 'This section does not exist').to.exist
         expect(results, 'No results whatsoever for this section').to.be.not.empty
@@ -75,9 +75,9 @@ describe('eDoujin Tests', () => {
 
     it('Testing search', async () => {
         const testSearch: SearchRequest = {
-            title: 'love',
+            title: 'the',
             parameters: {
-                includedTags: ['action']
+                includedTags: []
             }
         }
 
