@@ -1,32 +1,40 @@
-import { LanguageCode, SourceInfo, TagType } from "paperback-extensions-common";
-import { MangaStream } from '../MangaStream'
+/* eslint-disable linebreak-style */
+import {
+    LanguageCode,
+    SourceInfo,
+    ContentRating,
+    TagType
+} from 'paperback-extensions-common'
 
-const FLAMESCANS_DOMAIN = "https://flamescans.org"
+import {
+    MangaStream,
+    getExportVersion
+} from '../MangaStream'
 
-export const FlameScansInfo: SourceInfo = {
-    version: '1.0.2',
-    name: 'FlameScans',
-    description: 'Extension that pulls manga from FlameScans',
+const XCALIBRSCANS_DOMAIN = 'https://xcalibrscans.com'
+
+export const xCalibrScansInfo: SourceInfo = {
+    version: getExportVersion('0.0.0'),
+    name: 'xCalibrScans',
+    description: 'Extension that pulls manga from xCalibrScans',
     author: 'Netsky',
     authorWebsite: 'http://github.com/TheNetsky',
-    icon: "icon.png",
-    hentaiSource: false,
-    websiteBaseURL: FLAMESCANS_DOMAIN,
+    icon: 'icon.png',
+    contentRating: ContentRating.MATURE,
+    websiteBaseURL: XCALIBRSCANS_DOMAIN,
     sourceTags: [
         {
-            text: "Notifications",
+            text: 'Notifications',
             type: TagType.GREEN
         }
     ]
 }
 
-export class FlameScans extends MangaStream {
+export class xCalibrScans extends MangaStream {
     //FOR ALL THE SELECTIONS, PLEASE CHECK THE MangaSteam.ts FILE!!!
 
-    baseUrl: string = FLAMESCANS_DOMAIN
+    baseUrl: string = XCALIBRSCANS_DOMAIN
     languageCode: LanguageCode = LanguageCode.ENGLISH
-    hasAdvancedSearchPage: boolean = true
-    sourceTraversalPathName: string = 'series'
 
     //----MANGA DETAILS SELECTORS
     /*
@@ -45,15 +53,15 @@ export class FlameScans extends MangaStream {
     //Disabling some of these will cause some Home-Page tests to fail, edit these test files to match the setting.
     //Always be sure to test this in the app!
 
-    homescreen_PopularToday_enabled: boolean = true
+    override homescreen_PopularToday_enabled = true
 
-    homescreen_LatestUpdate_enabled: boolean = true
+    override homescreen_LatestUpdate_enabled = true
 
-    homescreen_NewManga_enabled: boolean = false
+    override homescreen_NewManga_enabled = false
 
-    homescreen_TopAllTime_enabled: boolean = true
-    homescreen_TopMonthly_enabled: boolean = true
-    homescreen_TopWeekly_enabled: boolean = true
+    override homescreen_TopAllTime_enabled = true
+    override homescreen_TopMonthly_enabled = true
+    override homescreen_TopWeekly_enabled = true
 
     /*
     ----TAG SELECTORS
@@ -69,4 +77,5 @@ export class FlameScans extends MangaStream {
     tags_selector_item: string = "li"
     tags_selector_label: string = "span"
     */
+
 }
